requirements.txt
boto3
pyyaml
mypy
types-PyYAML
black
pylint
pytest
